+++ src/sqlacodegen/generators.py
@@ -414,7 +414,10 @@ class TablesGenerator(CodeGenerator):
 
     def render_index(self, index: Index) -> str:
         extra_args = [repr(col.name) for col in index.columns]
-        kwargs = {}
+        kwargs = {
+            key: repr(value) if isinstance(value, str) else value
+            for key, value in sorted(index.kwargs.items())
+        }
         if index.unique:
             kwargs["unique"] = True